try to remove workspace files without executing chmod before; next try

+++ lib/common.sh
@@ -107,11 +107,13 @@ mustHaveCleanWorkspace() {
 removeWorkspace() {
     local workspace=$1
 
-    
-    execute rm -rf "${workspace}"
-    files_left=$(execute -n ls -R1|wc -l)
-    debug "files left after executing \"rm -rf ${workspace}\": ${files_left}"
-    if [[ ${files_left} -ge 1 ]] ; then
+    execute -i rm -rf "${workspace}"
+
+    # it may happen that some files could not be removed due to wrong access rights.
+    # if that's the case, change access rights and perform the rm command again.
+    if [[ -d ${workspace} ]] ; then
+        local files_left=$(execute -n ls -R1 ${workspace} | wc -l)
+        debug "files left after executing \"rm -rf ${workspace}\": ${files_left}"
         debug "changing access rights for left files"
         execute chmod -R u+w "${workspace}"
         execute rm -rf "${workspace}"
